refactor: split EtymologyCard into section components
Break the 588-line EtymologyCard into an orchestrator plus focused
section components under components/etymology-card/ (EntryHeader,
AncestrySection, StorySection, UsageSection, ModernUsageSection,
RelatedWordsSection, KinSection, SourcesSection, shared MobileSection).
Rendering is unchanged; markup and class names move verbatim. This
prepares the card for per-section progressive rendering in a later
branch.
